+++ contracts/babylon/src/multitest/suite.rs
@@ -160,6 +160,7 @@ impl SuiteBuilder {
                     consumer_description: Some("Test Consumer Description".to_string()),
                     ics20_channel_id: self.ics20_channel_id,
                     ibc_packet_timeout_days: None,
+                    destination_module: "btcstaking".to_string(),
                 },
                 &[],
                 "babylon",