+++ contracts/btc-finality/src/finality.rs
@@ -3,7 +3,7 @@ use crate::error::ContractError;
 use crate::state::config::{ADMIN, CONFIG, PARAMS};
 use crate::state::finality::{
     ensure_fp_has_power, get_power_table_at_height, BLOCKS, EVIDENCES, FP_BLOCK_SIGNER,
-    FP_POWER_TABLE, FP_START_HEIGHT, JAIL, NEXT_HEIGHT, SIGNATURES,
+    FP_POWER_TABLE, FP_START_HEIGHT, JAIL, NEXT_HEIGHT, REWARDS, SIGNATURES, TOTAL_PENDING_REWARDS,
 };
 use crate::state::public_randomness::{
     get_last_finalized_height, get_last_pub_rand_commit,
@@ -18,7 +18,7 @@ use btc_staking::msg::{FinalityProviderInfo, FinalityProvidersByTotalActiveSatsR
 use cosmwasm_std::Order::Ascending;
 use cosmwasm_std::{
     to_json_binary, Addr, DepsMut, Env, Event, MessageInfo, QuerierWrapper, Response, StdResult,
-    Storage, WasmMsg,
+    Storage, Uint128, WasmMsg,
 };
 use k256::schnorr::{signature::Verifier, Signature, VerifyingKey};
 use k256::sha2::{Digest, Sha256};
@@ -813,4 +813,53 @@ pub fn query_fps_by_total_active_sats(
     )?;
     let res: FinalityProvidersByTotalActiveSatsResponse = querier.query(&query)?;
     Ok(res.fps)
+}
+
+/// Distributes rewards to finality providers who are in the active set at `height`.
+pub fn distribute_rewards_fps(deps: &mut DepsMut, env: &Env) -> Result<(), ContractError> {
+    // Try to use the finality provider set at the previous height
+    let active_fps = get_power_table_at_height(deps.storage, env.block.height - 1)?;
+    // Short-circuit if there are no active finality providers
+    if active_fps.is_empty() {
+        return Ok(());
+    }
+    // Get the voting power of the active FPS
+    let total_voting_power = active_fps
+        .iter()
+        .map(|(_, power)| *power as u128)
+        .sum::<u128>();
+    // Short-circuit if the total voting power is zero
+    if total_voting_power == 0 {
+        return Ok(());
+    }
+    // Get the rewards to distribute (bank balance of the finality contract, minus previously / already distributed rewards
+    // (pending to be sent to Babylon on an epoch boundary))
+    let total_pending_rewards = TOTAL_PENDING_REWARDS.load(deps.storage)?;
+    let cfg = CONFIG.load(deps.storage)?;
+    let rewards_amount = deps
+        .querier
+        .query_balance(env.contract.address.clone(), cfg.denom)?
+        .amount
+        .saturating_sub(total_pending_rewards);
+    // Short-circuit if there are no rewards to distribute
+    if rewards_amount.is_zero() {
+        return Ok(());
+    }
+    // Compute the rewards for each active FP
+    let mut accumulated_rewards = Uint128::zero();
+    for (fp_btc_pk_hex, power) in active_fps {
+        let reward = (rewards_amount.u128() * power as u128) / total_voting_power;
+        let reward = Uint128::from(reward);
+        // Update the rewards for this FP
+        REWARDS.update(deps.storage, &fp_btc_pk_hex, |r| {
+            Ok::<Uint128, ContractError>(r.unwrap_or_default() + reward)
+        })?;
+        // Compute the total rewards
+        accumulated_rewards += reward;
+    }
+    // Update the total rewards
+    TOTAL_PENDING_REWARDS.update(deps.storage, |r| {
+        Ok::<Uint128, ContractError>(r + accumulated_rewards)
+    })?;
+    Ok(())
 }